feat(api): /v1 and /action/trigger
