+++ packages/server/lib/controllers/sync.controller.ts
@@ -1,36 +1,25 @@
-import tracer from 'dd-trace';
-
-import { getAccountUsageTracker } from '@nangohq/account-usage';
-import { OtlpSpan, defaultOperationExpiration, logContextGetter } from '@nangohq/logs';
+import { logContextGetter } from '@nangohq/logs';
 import { records as recordsService } from '@nangohq/records';
 import {
     NangoError,
     configService,
     connectionService,
     errorManager,
-    getActionOrModelByEndpoint,
     getSyncConfigRaw,
     getSyncs,
-    productTracking,
     syncCommandToOperation,
     verifyOwnership
 } from '@nangohq/shared';
-import { baseUrl, getHeaders, isCloud, isHosted, redactHeaders, truncateJson } from '@nangohq/utils';
+import { isHosted } from '@nangohq/utils';
 
-import { pubsub } from '../pubsub.js';
 import { getOrchestrator } from '../utils/utils.js';
-import { getPublicRecords } from './records/getRecords.js';
 
 import type { RequestLocals } from '../utils/express.js';
 import type { LogContextOrigin } from '@nangohq/logs';
-import type { HTTP_METHOD, Sync, SyncCommand } from '@nangohq/shared';
-import type { Span } from 'dd-trace';
+import type { Sync, SyncCommand } from '@nangohq/shared';
 import type { NextFunction, Request, Response } from 'express';
 
 const orchestrator = getOrchestrator();
-const accountUsageTracker = await getAccountUsageTracker();
-
-const actionPayloadAllowList = isCloud ? [662, 1760, 1920, 4530, 5166, 7157, 7359, 7696, 2981, 6254] : [];
 
 class SyncController {
     public async getSyncsByParams(req: Request, res: Response<any, Required<RequestLocals>>, next: NextFunction) {
@@ -84,260 +73,6 @@ class SyncController {
             }));
         } else {
             return syncs.map((sync) => ({ ...sync, record_count: null }));
-        }
-    }
-
-    public async actionOrModel(req: Request, res: Response<any, Required<RequestLocals>>, next: NextFunction) {
-        try {
-            const environmentId = res.locals['environment'].id;
-            const providerConfigKey = req.get('Provider-Config-Key') as string;
-            const connectionId = req.get('Connection-Id') as string;
-            const url = new URL(req.originalUrl, baseUrl);
-            const path = url.pathname.replace(/^\/v1\//, '/');
-            if (!connectionId) {
-                res.status(400).send({ error: 'Missing connection id' });
-
-                return;
-            }
-
-            if (!providerConfigKey) {
-                res.status(400).send({ error: 'Missing provider config key' });
-
-                return;
-            }
-            const { success, error, response: connection } = await connectionService.getConnection(connectionId, providerConfigKey, environmentId);
-
-            if (!success || !connection) {
-                errorManager.errResFromNangoErr(res, error);
-                return;
-            }
-
-            const { action, model } = await getActionOrModelByEndpoint(connection, req.method as HTTP_METHOD, path);
-            if (action) {
-                const input = req.body || req.params[1];
-                req.body = {};
-                req.body['action_name'] = action;
-                req.body['input'] = input;
-                await this.triggerAction(req, res, next);
-            } else if (model) {
-                Object.defineProperty(req, 'query', { ...Object.getOwnPropertyDescriptor(req, 'query'), value: req.query, writable: true });
-                req.query['model'] = model;
-                getPublicRecords(req, res, next);
-            } else {
-                res.status(404).send({ message: `Unknown endpoint '${req.method} ${path}'` });
-            }
-        } catch (err) {
-            next(err);
-        }
-    }
-
-    public async triggerAction(req: Request, res: Response<any, Required<RequestLocals>>, next: NextFunction) {
-        const active = tracer.scope().active();
-        const span = tracer.startSpan('server.sync.triggerAction', {
-            childOf: active as Span
-        });
-
-        const { input, action_name } = req.body;
-        const { account, environment, plan } = res.locals;
-
-        const environmentId = environment.id;
-        const connectionId = req.get('Connection-Id');
-        const providerConfigKey = req.get('Provider-Config-Key');
-        const async = req.get('X-Async') === 'true';
-        const retryMaxHeaderName = 'X-Max-Retries';
-        const retryMax = parseInt(req.get(retryMaxHeaderName) || '0');
-        let logCtx: LogContextOrigin | undefined;
-        try {
-            if (!action_name || typeof action_name !== 'string') {
-                res.status(400).send({ error: 'Missing action name' });
-
-                span.finish();
-                return;
-            }
-
-            if (!connectionId) {
-                res.status(400).send({ error: 'Missing connection id' });
-
-                span.finish();
-                return;
-            }
-
-            if (!providerConfigKey) {
-                res.status(400).send({ error: 'Missing provider config key' });
-
-                span.finish();
-                return;
-            }
-
-            if (async && (retryMax < 0 || retryMax > 5)) {
-                res.status(400).send({ error: `${retryMaxHeaderName} must be between 0 and 5` });
-
-                span.finish();
-                return;
-            }
-
-            const { success, error, response: connection } = await connectionService.getConnection(connectionId, providerConfigKey, environmentId);
-
-            if (!success || !connection) {
-                errorManager.errResFromNangoErr(res, error);
-
-                span.finish();
-                return;
-            }
-
-            const provider = await configService.getProviderConfig(providerConfigKey, environmentId);
-            if (!provider) {
-                res.status(404).json({ error: { code: 'not_found' } });
-                return;
-            }
-
-            const syncConfig = await getSyncConfigRaw({ environmentId, config_id: provider.id!, name: action_name, isAction: true });
-            if (!syncConfig) {
-                res.status(404).json({ error: { code: 'not_found' } });
-                return;
-            }
-
-            if (!syncConfig.enabled) {
-                res.status(404).json({ error: { code: 'disabled_resource', message: 'The action is disabled' } });
-                return;
-            }
-
-            span.setTag('nango.actionName', action_name)
-                .setTag('nango.connectionId', connectionId)
-                .setTag('nango.environmentId', environmentId)
-                .setTag('nango.providerConfigKey', providerConfigKey);
-
-            logCtx = await logContextGetter.create(
-                { operation: { type: 'action', action: 'run' }, expiresAt: defaultOperationExpiration.action() },
-                {
-                    account,
-                    environment,
-                    integration: { id: provider.id!, name: connection.provider_config_key, provider: provider.provider },
-                    connection: { id: connection.id, name: connection.connection_id },
-                    syncConfig: { id: syncConfig.id, name: syncConfig.sync_name },
-                    meta: truncateJson({ input })
-                }
-            );
-            logCtx.attachSpan(new OtlpSpan(logCtx.operation));
-
-            if (plan && (await accountUsageTracker.shouldCapUsage(plan, 'actions'))) {
-                const message =
-                    'Your monthly limit for action executions has been reached. No actions will run until you upgrade your account or the limit resets.';
-                void logCtx.error(message, {
-                    usage: await accountUsageTracker.getUsage({ accountId: account.id, metric: 'actions' }),
-                    limit: accountUsageTracker.getLimit(plan, 'actions')
-                });
-
-                productTracking.track({ name: 'server:resource_capped:action_triggered', team: account });
-
-                res.status(400).send({
-                    error: {
-                        code: 'resource_capped',
-                        message
-                    }
-                });
-
-                span.setTag('nango.usageLimitExceeded', true);
-                await logCtx.failed();
-                span.finish();
-                return;
-            }
-
-            const actionResponse = await getOrchestrator().triggerAction({
-                accountId: account.id,
-                connection,
-                actionName: action_name,
-                input,
-                async,
-                retryMax,
-                logCtx
-            });
-
-            if (actionResponse.isOk()) {
-                span.finish();
-
-                const payloadSize = Buffer.byteLength(JSON.stringify(actionResponse.value), 'utf8');
-                const payloadSizeInMb = payloadSize / (1024 * 1024);
-
-                if (payloadSizeInMb > 2) {
-                    if (actionPayloadAllowList.includes(account.id)) {
-                        void logCtx.warn(
-                            `The action payload is larger than 2 MB at ${payloadSizeInMb}. The usage of an action for an output this large wil soon be deprecated. It is recommended to use the nango proxy directly for such operations. See the proxy docs: https://docs.nango.dev/guides/proxy-requests#proxy-requests.`,
-                            {
-                                payloadSize,
-                                actionName: action_name,
-                                connectionId: connection.id,
-                                environmentId: environment.id
-                            }
-                        );
-                    }
-
-                    if (!actionPayloadAllowList.includes(account.id)) {
-                        void logCtx.error('Action payload is larger than 2 MB and will be blocked in the near future. Please use the proxy', {
-                            payloadSize,
-                            actionName: action_name,
-                            connectionId: connection.id,
-                            environmentId: environment.id
-                        });
-                    }
-                }
-
-                if ('statusUrl' in actionResponse.value) {
-                    res.status(202).location(actionResponse.value.statusUrl).json(actionResponse.value);
-                } else {
-                    res.status(200).json(actionResponse.value.data);
-                }
-
-                void pubsub.publisher.publish({
-                    subject: 'usage',
-                    type: 'usage.actions',
-                    idempotencyKey: logCtx.id,
-                    payload: {
-                        value: 1,
-                        properties: {
-                            accountId: account.id,
-                            connectionId: connection.id,
-                            environmentId: connection.environment_id,
-                            providerConfigKey,
-                            actionName: action_name
-                        }
-                    }
-                });
-
-                return;
-            } else {
-                span.setTag('nango.error', actionResponse.error);
-                await logCtx.failed();
-
-                errorManager.errResFromNangoErr(res, actionResponse.error);
-
-                span.finish();
-                return;
-            }
-        } catch (err) {
-            span.setTag('nango.error', err);
-            span.finish();
-            if (logCtx) {
-                void logCtx.error('Failed to run action', { error: err });
-                await logCtx.failed();
-            }
-
-            next(err);
-        } finally {
-            const reqHeaders = getHeaders(req.headers);
-            reqHeaders['authorization'] = 'REDACTED';
-            const responseHeaders = getHeaders(res.getHeaders());
-            await logCtx?.enrichOperation({
-                request: {
-                    url: `${req.protocol}://${req.get('host')}${req.originalUrl}`,
-                    method: req.method,
-                    headers: redactHeaders({ headers: reqHeaders })
-                },
-                response: {
-                    code: res.statusCode,
-                    headers: redactHeaders({ headers: responseHeaders })
-                }
-            });
         }
     }
