+++ packages/types/lib/action/api.ts
@@ -16,3 +16,32 @@ export type GetAsyncActionResult = Endpoint<{
     // TODO: fix Endpoint definition to support any json value
     Success: Record<string, any>;
 }>;
+
+export type PostPublicTriggerAction = Endpoint<{
+    Method: 'POST';
+    Path: '/action/trigger';
+    Body: {
+        action_name: string;
+        input: any;
+    };
+    Headers: {
+        'provider-config-key': string;
+        'connection-id': string;
+        'x-async'?: boolean;
+        'x-max-retries'?: number;
+    };
+    Success: any;
+}>;
+
+export type GetPublicV1 = Endpoint<{
+    Method: 'GET';
+    Path: `/v1/:path`;
+    Params: any;
+    Body: any;
+    Querystring: any;
+    Headers: {
+        'provider-config-key': string;
+        'connection-id': string;
+    };
+    Success: any;
+}>;
